# Optimize convolution and BN in DenseNet block