I have added the convolution optimizations, and tagged a loop in batch normalization to be parallelized.

The convolution optimizations are : loop interchange, parallelize, split and vectorize. I have tried to tune the parameters according to the different sizes of the input.

For BN, I should see how to optimize it better.

Here are the results that I have, the input and output channels are always respectively 128 and 32. **N** represents the width and height of the input. As you see, for N = 7, Intel MKL code outperforms TIramisu code, I should investigate this.

-----------------------------------

BATCH_SIZE = 100
N = 56
MKL : 3366.581 ms
Tiramisu : 2075.63 ms

-----------------------------------

BATCH_SIZE = 32
N = 56
MKL : 1056.836 ms
Tiramisu : 666.363 ms

-----------------------------------

BATCH_SIZE = 8
N = 56
MKL : 264.575 ms
Tiramisu : 155.588 ms

-----------------------------------

BATCH_SIZE = 100
N = 28
MKL : 809.433 ms
Tiramisu : 555.059 ms

-----------------------------------

BATCH_SIZE = 32
N = 28
MKL : 256.138 ms
Tiramisu : 171.093 ms

-----------------------------------

BATCH_SIZE = 8
N = 28
MKL : 64.023 ms
Tiramisu : 42.788 ms

-----------------------------------

BATCH_SIZE = 100
N = 14
MKL : 186.153 ms
Tiramisu : 144.713 ms

-----------------------------------

BATCH_SIZE = 32
N = 14
MKL : 62.888 ms
Tiramisu : 48.116 ms

-----------------------------------

BATCH_SIZE = 8
N = 14
MKL : 15.047 ms
Tiramisu : 12.055 ms

-----------------------------------

BATCH_SIZE = 100
N = 7
MKL : 47.477 ms
Tiramisu : 50.560 ms

-----------------------------------

BATCH_SIZE = 32
N = 7
MKL : 15.301 ms
Tiramisu : 16.825 ms

-----------------------------------

BATCH_SIZE = 8
N = 7
MKL : 1.798 ms
Tiramisu : 4.253 ms