@rbaghdadi 2019-06-05
Great! Thanks Kamel, the numbers look interesting.

Can you create a google spreadsheet file to log all of these numbers along with the machine characteristics, data sizes, ... so that we can reproduce them?

For small data sizes, you might want to change the schedule. In our reference convolution implementation, we have certain optimizations that are disabled for small data sizes, are you doing the same?

Also for very small sizes you might want to disable parallelism, and tiling since parallelization and tiling have an overhead are only useful if the data is big enough.