@abdouskamel 2019-06-05
@rbaghdadi For small sizes, in convolution, I have left only vectorization of the innermost loop and parallelization over the batch size. I thought that I could remove the parallelization, but after testing, I found that it actually decreases execution time significantly, even for small sizes.

I will do for the spreadsheet.